Changed some names
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@2083 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserver/source/security.cpp
@@ -62,7 +62,7 @@ void SecurityToken::reset()
   permission2    = 0;
   authType      = 0;
   authTypeLen       = 0;
-  throttlingRate = (u_long)-1;
+  throttlingRate = (int)-1;
 }
 
 /*!
